requirements.txt
# Nexus AI Test Agent - Dependencies

# Core
pyyaml>=6.0
pydantic>=2.0

# API Parsing
openapi-spec-validator>=0.7.0

# Testing
pytest>=7.0
pytest-html>=4.0
pytest-timeout>=2.2

# Web (Sprint 2)
flask>=3.0
flask-socketio>=5.3

# CLI output
rich>=13.0

# HTTP requests for testing
requests>=2.31
